fix: 장소 입력 안되는 버그

+++ src/components/common/TodoListItem/index.tsx
@@ -63,8 +63,10 @@ const TodoListItem: React.FC<TodoListItemProps> = ({
   }, []);
 
   useEffect(() => {
-    const deleteTodoHandler = () => {
-      deleteTodo(id);
+    const deleteTodoHandler = (e: KeyboardEvent) => {
+      if (e.key === 'Backspace' || e.key === 'Delete') {
+        deleteTodo(id);
+      }
     };
 
     if (isFocused) {
